refactor: split large files + drop narrative comments for 100/100 aislop
- SkinMetrics (570 -> 280 lines): latency buckets moved to LatencyHistogram,
  rendering moved to MetricsFormat (public utility)
- SkinCommand (624 -> 340 lines): /skin metrics tree moved to
  SkinMetricsCommand; set/source/clear execution + rate limit + debounce +
  completion handling moved to SkinActionCommand (shared state moved with it;
  SkinRestorer + gametests updated)
- SkinRefreshHandler.task() (94 -> 28 lines): extracted mutateProfile /
  recordSaveEnqueue / recordObserverBroadcast / recordCascade
- Removed 24 narrative javadoc blocks from the gametest (test names are
  self-descriptive); kept why-comments (UPDATE_DISPLAY_NAME bug rationale,
  drain latch race, KEEP_ALL_DATA convention)
- SkinRefreshHandler helpers made public (now called cross-package)
- All 257 unit tests + 27 GameTests pass unchanged; zero new dependencies

Aislop: 96/100 -> 100/100
