+++ src/main/java/levosilimo/everlastingskins/metrics/SkinMetrics.java
@@ -1,6 +1,5 @@
 package levosilimo.everlastingskins.metrics;
 
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -14,21 +13,10 @@ import java.util.concurrent.atomic.LongAdder;
 /**
  * In-process metrics for the EverlastingSkins /skin refresh flow.
  * Zero external dependencies, lock-free writes from any thread.
- *
- * Latency tracking uses fixed buckets (not HdrHistogram):
- * 1us, 5us, 10us, 50us, 100us, 500us, 1ms, 5ms, 10ms, 50ms, 100ms,
- * 250ms, 500ms, 1s, 5s, 10s, plus an overflow bucket for everything above.
- * p50/p95/p99 are computed by walking the cumulative distribution.
+ * Latency distributions live in {@link LatencyHistogram} fixed buckets;
+ * rendering lives in {@link MetricsFormat}.
  */
 public final class SkinMetrics {
-
-    private static final long[] LATENCY_BUCKETS_US = {
-            1, 5, 10, 50, 100, 500, 1_000, 5_000, 10_000, 50_000,
-            100_000, 250_000, 500_000, 1_000_000, 5_000_000, 10_000_000
-    };
-
-    /** A task phase exceeding this threshold counts as a tick spike. */
-    private static final long TICK_SPIKE_THRESHOLD_NANOS = 50_000_000;
 
     public static final SkinMetrics INSTANCE = new SkinMetrics();
 
@@ -62,13 +50,13 @@ public final class SkinMetrics {
 
     private final ConcurrentHashMap<String, LongAdder> ioFailuresByType = new ConcurrentHashMap<>();
 
-    private final LongAdder[] fetchLatencyBuckets = new LongAdder[LATENCY_BUCKETS_US.length + 1];
-    private final LongAdder[] saveEnqueueLatencyBuckets = new LongAdder[LATENCY_BUCKETS_US.length + 1];
-    private final LongAdder[] saveDiskLatencyBuckets = new LongAdder[LATENCY_BUCKETS_US.length + 1];
-    private final LongAdder[] broadcastLatencyBuckets = new LongAdder[LATENCY_BUCKETS_US.length + 1];
-    private final LongAdder[] commandTotalLatencyBuckets = new LongAdder[LATENCY_BUCKETS_US.length + 1];
-    private final LongAdder[] taskDurationLatencyBuckets = new LongAdder[LATENCY_BUCKETS_US.length + 1];
-    private final LongAdder[] tickSpikeLatencyBuckets = new LongAdder[LATENCY_BUCKETS_US.length + 1];
+    private final LatencyHistogram fetchLatency = new LatencyHistogram();
+    private final LatencyHistogram saveEnqueueLatency = new LatencyHistogram();
+    private final LatencyHistogram saveDiskLatency = new LatencyHistogram();
+    private final LatencyHistogram broadcastLatency = new LatencyHistogram();
+    private final LatencyHistogram commandTotalLatency = new LatencyHistogram();
+    private final LatencyHistogram taskDurationLatency = new LatencyHistogram();
+    private final LatencyHistogram tickSpikeLatency = new LatencyHistogram();
 
     private final AtomicInteger pendingAsyncWrites = new AtomicInteger();
     private final LongAdder onlinePlayers = new LongAdder();
@@ -78,44 +66,27 @@ public final class SkinMetrics {
 
     /** Package-private for tests; production uses {@link #INSTANCE}. */
     SkinMetrics() {
-        for (int i = 0; i <= LATENCY_BUCKETS_US.length; i++) {
-            fetchLatencyBuckets[i] = new LongAdder();
-            saveEnqueueLatencyBuckets[i] = new LongAdder();
-            saveDiskLatencyBuckets[i] = new LongAdder();
-            broadcastLatencyBuckets[i] = new LongAdder();
-            commandTotalLatencyBuckets[i] = new LongAdder();
-            taskDurationLatencyBuckets[i] = new LongAdder();
-            tickSpikeLatencyBuckets[i] = new LongAdder();
-        }
     }
 
     public void recordRefreshStarted(UUID player) {
         refreshesInitiated.increment();
     }
 
-    /**
-     * Records a completed refresh. fetch/save/broadcast latencies are recorded
-     * into their own histograms when non-zero (the save and broadcast phases
-     * run asynchronously from the command completion, so callers pass what
-     * they measured); the command-span latency (startNanos to now) is recorded
-     * into the commandTotal histogram.
-     */
     public void recordRefreshCompleted(UUID player, long startNanos, long fetchNanos, long saveNanos, long broadcastNanos) {
         refreshesCompleted.increment();
         PlayerMetrics pm = perPlayer.computeIfAbsent(player, k -> new PlayerMetrics());
         pm.refreshCount.increment();
         pm.lastRefreshAtMs.set(System.currentTimeMillis());
-        recordLatency(fetchLatencyBuckets, fetchNanos);
-        recordLatency(saveEnqueueLatencyBuckets, saveNanos);
-        recordLatency(broadcastLatencyBuckets, broadcastNanos);
-        recordLatency(commandTotalLatencyBuckets, System.nanoTime() - startNanos);
+        fetchLatency.record(fetchNanos);
+        saveEnqueueLatency.record(saveNanos);
+        broadcastLatency.record(broadcastNanos);
+        commandTotalLatency.record(System.nanoTime() - startNanos);
     }
 
     public void recordRefreshFailed(UUID player) {
         refreshesFailed.increment();
     }
 
-    /** Fetch timed out (10s), distinct from a provider failure. */
     public void recordTimedOut(UUID player) {
         refreshesTimedOut.increment();
     }
@@ -128,7 +99,6 @@ public final class SkinMetrics {
         refreshesSkipped.increment();
     }
 
-    /** Skipped because the stored skin's source already matches the request. */
     public void recordRefreshSkippedStored(UUID player) {
         refreshesSkippedStored.increment();
     }
@@ -142,53 +112,46 @@ public final class SkinMetrics {
         bytesWritten.add(bytes);
     }
 
-    /** Time between submit and the writer thread picking the write up. */
     public void recordSaveEnqueueLatency(long nanos) {
-        recordLatency(saveEnqueueLatencyBuckets, nanos);
+        saveEnqueueLatency.record(nanos);
     }
 
-    /** Actual disk write (write + fsync + atomic rename), measured on the writer thread. */
     public void recordSaveDiskLatency(long nanos) {
-        recordLatency(saveDiskLatencyBuckets, nanos);
+        saveDiskLatency.record(nanos);
     }
 
     public void recordBroadcastLatency(long nanos) {
-        recordLatency(broadcastLatencyBuckets, nanos);
+        broadcastLatency.record(nanos);
     }
 
-    /** Records the full task() duration into the task-duration histogram. */
     public void recordTaskDuration(long nanos) {
-        recordLatency(taskDurationLatencyBuckets, nanos);
+        taskDurationLatency.record(nanos);
     }
 
-    /** Consumes the per-connection byte deltas measured around a refresh. */
     public void recordNetworkDelta(long outDelta, long inDelta) {
         if (outDelta > 0) netBytesWrittenOut.add(outDelta);
         if (inDelta > 0) netBytesReadIn.add(inDelta);
     }
 
-    /** A task exceeded the 50ms spike threshold (generic count + histogram). */
     public void recordTickSpike(long nanos) {
         tickSpikes.increment();
-        recordLatency(tickSpikeLatencyBuckets, nanos);
+        tickSpikeLatency.record(nanos);
     }
 
-    /** A broadcast-phase spike (the REMOVE + ADD_PLAYER fan-out). */
     public void recordSpikeBroadcast(long nanos) {
         if (nanos >= TICK_SPIKE_THRESHOLD_NANOS) tickSpikesBroadcast.increment();
     }
 
-    /** A respawn-cascade-phase spike (per-target packet sequence). */
     public void recordSpikeCascade(long nanos) {
         if (nanos >= TICK_SPIKE_THRESHOLD_NANOS) tickSpikesCascade.increment();
     }
 
-    /** A save-submit-phase spike (coalescing enqueue). */
     public void recordSpikeSaveEnqueue(long nanos) {
         if (nanos >= TICK_SPIKE_THRESHOLD_NANOS) tickSpikesSaveEnqueue.increment();
     }
 
-    /** Provider HTTP status code observed by the HTTP layer. */
+    private static final long TICK_SPIKE_THRESHOLD_NANOS = 50_000_000;
+
     public void recordProviderStatus(int statusCode) {
         if (statusCode == 429) {
             providerHttp429.increment();
@@ -221,12 +184,10 @@ public final class SkinMetrics {
         pendingAsyncWrites.decrementAndGet();
     }
 
-    /** A submission was superseded by a newer payload for the same UUID. */
     public void recordSaveCoalesced() {
         savesCoalesced.increment();
     }
 
-    /** An actual disk write happened (write + fsync + rename). */
     public void recordRealWrite() {
         realWrites.increment();
     }
@@ -235,7 +196,6 @@ public final class SkinMetrics {
         ioFailures.increment();
     }
 
-    /** IOException with type attribution (NoSpaceLeft, AccessDenied, ...). */
     public void recordIoFailure(Throwable t) {
         ioFailures.increment();
         String type = t != null ? t.getClass().getSimpleName() : "unknown";
@@ -268,7 +228,6 @@ public final class SkinMetrics {
         return removed[0];
     }
 
-    /** Zeroes every counter and clears per-player state. */
     public void reset() {
         refreshesInitiated.reset();
         refreshesCompleted.reset();
@@ -298,71 +257,17 @@ public final class SkinMetrics {
         cacheHits.reset();
         cacheMisses.reset();
         ioFailuresByType.clear();
-        for (int i = 0; i <= LATENCY_BUCKETS_US.length; i++) {
-            fetchLatencyBuckets[i].reset();
-            saveEnqueueLatencyBuckets[i].reset();
-            saveDiskLatencyBuckets[i].reset();
-            broadcastLatencyBuckets[i].reset();
-            commandTotalLatencyBuckets[i].reset();
-            taskDurationLatencyBuckets[i].reset();
-            tickSpikeLatencyBuckets[i].reset();
-        }
+        fetchLatency.reset();
+        saveEnqueueLatency.reset();
+        saveDiskLatency.reset();
+        broadcastLatency.reset();
+        commandTotalLatency.reset();
+        taskDurationLatency.reset();
+        tickSpikeLatency.reset();
         pendingAsyncWrites.set(0);
         onlinePlayers.reset();
         perPlayer.clear();
         startedAtMs.set(System.currentTimeMillis());
-    }
-
-    private static void recordLatency(LongAdder[] buckets, long nanos) {
-        if (nanos <= 0) return;
-        long us = Math.max(1, nanos / 1000);
-        int idx = Arrays.binarySearch(LATENCY_BUCKETS_US, us);
-        if (idx < 0) idx = -idx - 1;
-        buckets[idx].increment();
-    }
-
-    /**
-     * Computes p50/p95/p99 (in microseconds) and the highest non-empty bucket
-     * bound by walking the cumulative distribution. Values land on the bucket
-     * bound they fall into; samples beyond the last bound report as
-     * {@code maxBound} (the 10s+ overflow bucket).
-     */
-    static Map<String, Long> histogramPercentiles(LongAdder[] buckets, long[] bucketBoundsUs) {
-        Map<String, Long> result = new LinkedHashMap<>();
-        long total = 0;
-        for (LongAdder bucket : buckets) {
-            total += bucket.sum();
-        }
-        if (total == 0) {
-            result.put("p50", 0L);
-            result.put("p95", 0L);
-            result.put("p99", 0L);
-            result.put("max", 0L);
-            return result;
-        }
-        long cumulative = 0;
-        long maxBound = 0;
-        long p50 = 0;
-        long p95 = 0;
-        long p99 = 0;
-        long p50Target = Math.round(total * 0.50);
-        long p95Target = Math.round(total * 0.95);
-        long p99Target = Math.round(total * 0.99);
-        for (int i = 0; i < buckets.length; i++) {
-            long count = buckets[i].sum();
-            if (count == 0) continue;
-            cumulative += count;
-            long bound = i < bucketBoundsUs.length ? bucketBoundsUs[i] : bucketBoundsUs[bucketBoundsUs.length - 1];
-            maxBound = bound;
-            if (p50 == 0 && cumulative >= p50Target) p50 = bound;
-            if (p95 == 0 && cumulative >= p95Target) p95 = bound;
-            if (p99 == 0 && cumulative >= p99Target) p99 = bound;
-        }
-        result.put("p50", p50);
-        result.put("p95", p95);
-        result.put("p99", p99);
-        result.put("max", maxBound);
-        return result;
     }
 
     public Snapshot snapshot() {
@@ -380,13 +285,10 @@ public final class SkinMetrics {
                 providerHttp429.sum(), providerHttp5xx.sum(), providerHttp4xxOther.sum(),
                 providerExceptions.sum(), cacheHits.sum(), cacheMisses.sum(),
                 ioFailuresByType(),
-                histogramPercentiles(fetchLatencyBuckets, LATENCY_BUCKETS_US),
-                histogramPercentiles(saveEnqueueLatencyBuckets, LATENCY_BUCKETS_US),
-                histogramPercentiles(saveDiskLatencyBuckets, LATENCY_BUCKETS_US),
-                histogramPercentiles(broadcastLatencyBuckets, LATENCY_BUCKETS_US),
-                histogramPercentiles(commandTotalLatencyBuckets, LATENCY_BUCKETS_US),
-                histogramPercentiles(taskDurationLatencyBuckets, LATENCY_BUCKETS_US),
-                histogramPercentiles(tickSpikeLatencyBuckets, LATENCY_BUCKETS_US),
+                fetchLatency.percentiles(), saveEnqueueLatency.percentiles(),
+                saveDiskLatency.percentiles(), broadcastLatency.percentiles(),
+                commandTotalLatency.percentiles(), taskDurationLatency.percentiles(),
+                tickSpikeLatency.percentiles(),
                 snapshotPerPlayer());
     }
 
@@ -402,142 +304,6 @@ public final class SkinMetrics {
         Map<UUID, PlayerSnapshot> map = new LinkedHashMap<>();
         perPlayer.forEach((uuid, pm) -> map.put(uuid, new PlayerSnapshot(pm.refreshCount.sum(), pm.lastRefreshAtMs.get())));
         return map;
-    }
-
-    /** Formats the snapshot as the human-readable /skin metrics output. */
-    public String formatHuman(Snapshot s) {
-        StringBuilder sb = new StringBuilder();
-        sb.append("EverlastingSkins Metrics (uptime ").append(formatUptime(s.uptimeMs())).append(")\n");
-        sb.append("  refreshes: ").append(s.refreshesInitiated()).append(" initiated, ")
-                .append(s.refreshesCompleted()).append(" completed, ")
-                .append(s.refreshesFailed()).append(" failed, ")
-                .append(s.refreshesTimedOut()).append(" timed out\n");
-        sb.append("  skipped: ").append(s.refreshesSkipped()).append(" (identical), ")
-                .append(s.refreshesSkippedStored()).append(" (stored source), debounced: ")
-                .append(s.refreshesDebounced()).append(", rate-limited: ").append(s.refreshesRateLimited()).append('\n');
-        sb.append("  broadcasts: ").append(s.broadcastsSent()).append(" (").append(formatBytes(s.bytesWritten())).append(" written)\n");
-        sb.append("  saves: ").append(s.savesSubmitted()).append(" submitted, ").append(s.savesCompleted())
-                .append(" completed, ").append(s.savesCoalesced()).append(" coalesced, ")
-                .append(s.realWrites()).append(" real writes, ").append(s.ioFailures()).append(" failed, ")
-                .append(s.pendingAsyncWrites()).append(" pending\n");
-        sb.append("  io failures by type: ").append(formatIoFailures(s.ioFailuresByType())).append('\n');
-        sb.append("  network: ").append(formatBytes(s.netBytesWrittenOut())).append(" out, ")
-                .append(formatBytes(s.netBytesReadIn())).append(" in (per-connection)\n");
-        sb.append("  provider: ").append(s.providerHttp429()).append("x429, ")
-                .append(s.providerHttp5xx()).append("x5xx, ")
-                .append(s.providerHttp4xxOther()).append("x4xx, ")
-                .append(s.providerExceptions()).append(" exceptions\n");
-        sb.append("  cache: ").append(s.cacheHits()).append(" hits, ")
-                .append(s.cacheMisses()).append(" misses\n");
-        sb.append("  tick spikes: ").append(s.tickSpikes()).append(" (broadcast ")
-                .append(s.tickSpikesBroadcast()).append(", cascade ").append(s.tickSpikesCascade())
-                .append(", save-enqueue ").append(s.tickSpikesSaveEnqueue()).append(")\n");
-        sb.append("  online players: ").append(s.onlinePlayers()).append('\n');
-        sb.append("  latencies (ms):\n");
-        sb.append("    fetch:        ").append(formatPercentiles(s.fetchPercentiles())).append('\n');
-        sb.append("    save enq:     ").append(formatPercentiles(s.saveEnqueuePercentiles())).append('\n');
-        sb.append("    save disk:    ").append(formatPercentiles(s.saveDiskPercentiles())).append('\n');
-        sb.append("    broadcast:    ").append(formatPercentiles(s.broadcastPercentiles())).append('\n');
-        sb.append("    command total:").append(formatPercentiles(s.commandTotalPercentiles())).append('\n');
-        sb.append("    task duration:").append(formatPercentiles(s.taskDurationPercentiles())).append('\n');
-        sb.append("    tick spike:   ").append(formatPercentiles(s.tickSpikePercentiles()));
-        return sb.toString();
-    }
-
-    /** Formats the snapshot as a single JSON object. */
-    public String formatJson(Snapshot s) {
-        StringBuilder sb = new StringBuilder();
-        sb.append("{\"uptimeMs\":").append(s.uptimeMs());
-        sb.append(",\"refreshes\":{\"initiated\":").append(s.refreshesInitiated())
-                .append(",\"completed\":").append(s.refreshesCompleted())
-                .append(",\"failed\":").append(s.refreshesFailed())
-                .append(",\"timedOut\":").append(s.refreshesTimedOut())
-                .append(",\"skipped\":").append(s.refreshesSkipped())
-                .append(",\"skippedStored\":").append(s.refreshesSkippedStored())
-                .append(",\"debounced\":").append(s.refreshesDebounced())
-                .append(",\"rateLimited\":").append(s.refreshesRateLimited()).append('}');
-        sb.append(",\"broadcasts\":{\"sent\":").append(s.broadcastsSent())
-                .append(",\"bytesWritten\":").append(s.bytesWritten()).append('}');
-        sb.append(",\"saves\":{\"submitted\":").append(s.savesSubmitted())
-                .append(",\"completed\":").append(s.savesCompleted())
-                .append(",\"coalesced\":").append(s.savesCoalesced())
-                .append(",\"realWrites\":").append(s.realWrites())
-                .append(",\"ioFailures\":").append(s.ioFailures())
-                .append(",\"pending\":").append(s.pendingAsyncWrites()).append('}');
-        sb.append(",\"ioFailuresByType\":{");
-        boolean firstType = true;
-        for (Map.Entry<String, Long> e : s.ioFailuresByType().entrySet()) {
-            if (!firstType) sb.append(',');
-            firstType = false;
-            sb.append('"').append(e.getKey()).append("\":").append(e.getValue());
-        }
-        sb.append('}');
-        sb.append(",\"networkBytes\":{\"out\":").append(s.netBytesWrittenOut())
-                .append(",\"in\":").append(s.netBytesReadIn()).append('}');
-        sb.append(",\"provider\":{\"http429\":").append(s.providerHttp429())
-                .append(",\"http5xx\":").append(s.providerHttp5xx())
-                .append(",\"http4xxOther\":").append(s.providerHttp4xxOther())
-                .append(",\"exceptions\":").append(s.providerExceptions()).append('}');
-        sb.append(",\"cache\":{\"hits\":").append(s.cacheHits())
-                .append(",\"misses\":").append(s.cacheMisses()).append('}');
-        sb.append(",\"tickSpikes\":{\"total\":").append(s.tickSpikes())
-                .append(",\"broadcast\":").append(s.tickSpikesBroadcast())
-                .append(",\"cascade\":").append(s.tickSpikesCascade())
-                .append(",\"saveEnqueue\":").append(s.tickSpikesSaveEnqueue()).append('}');
-        sb.append(",\"onlinePlayers\":").append(s.onlinePlayers());
-        sb.append(",\"latenciesMs\":{")
-                .append("\"fetch\":").append(jsonPercentiles(s.fetchPercentiles()))
-                .append(",\"saveEnqueue\":").append(jsonPercentiles(s.saveEnqueuePercentiles()))
-                .append(",\"saveDisk\":").append(jsonPercentiles(s.saveDiskPercentiles()))
-                .append(",\"broadcast\":").append(jsonPercentiles(s.broadcastPercentiles()))
-                .append(",\"commandTotal\":").append(jsonPercentiles(s.commandTotalPercentiles()))
-                .append(",\"taskDuration\":").append(jsonPercentiles(s.taskDurationPercentiles()))
-                .append(",\"tickSpike\":").append(jsonPercentiles(s.tickSpikePercentiles())).append('}');
-        sb.append(",\"players\":[");
-        boolean first = true;
-        for (Map.Entry<UUID, PlayerSnapshot> e : s.perPlayer().entrySet()) {
-            if (!first) sb.append(',');
-            first = false;
-            sb.append("{\"uuid\":\"").append(e.getKey()).append("\",\"refreshCount\":")
-                    .append(e.getValue().refreshCount()).append(",\"lastRefreshAtMs\":")
-                    .append(e.getValue().lastRefreshAtMs()).append('}');
-        }
-        sb.append("]}");
-        return sb.toString();
-    }
-
-    private static String formatIoFailures(Map<String, Long> byType) {
-        if (byType.isEmpty()) return "none";
-        StringBuilder sb = new StringBuilder();
-        byType.forEach((type, count) -> sb.append(type).append(": ").append(count).append(", "));
-        return sb.substring(0, sb.length() - 2);
-    }
-
-    private static String jsonPercentiles(Map<String, Long> p) {
-        return "{\"p50\":" + p.get("p50") + ",\"p95\":" + p.get("p95")
-                + ",\"p99\":" + p.get("p99") + ",\"max\":" + p.get("max") + '}';
-    }
-
-    private static String formatPercentiles(Map<String, Long> p) {
-        return "p50=" + usToMs(p.get("p50")) + ", p95=" + usToMs(p.get("p95"))
-                + ", p99=" + usToMs(p.get("p99")) + ", max=" + usToMs(p.get("max"));
-    }
-
-    private static long usToMs(long us) {
-        return us / 1000;
-    }
-
-    private static String formatUptime(long ms) {
-        long totalMin = ms / 60_000;
-        long hours = totalMin / 60;
-        long minutes = totalMin % 60;
-        return hours + "h " + minutes + "m";
-    }
-
-    private static String formatBytes(long bytes) {
-        if (bytes < 1024) return bytes + " B";
-        if (bytes < 1024 * 1024) return (bytes / 1024) + " KB";
-        return (bytes / (1024 * 1024)) + " MB";
     }
 
     public record Snapshot(
